Complication Dagger Component introduction.
This changelist introduces a dagger component dedicated to
complications. This component will house complication-specific
entities as complications are separated from the dream overlay.
A factory for the component is provided to DreamOverlayService
for constructing complication dependencies in the near term.
Eventually, DreamOverlayService will be able to accept
complication components specified elsewhere to interact with
complications in other surfaces.

Test: atest DreamOverlayServiceTest
Bug: 261781069
Change-Id: I581f480a402afc2dbc1d052fd07ec152e9c0ea7b

Former-commit-id: 879f1bf42c45d80ea2e377ffde66d0f188843978