Remove InternalInsetsType from InsetsSourceConsumer
This CL uses the ID to identify an insets source instead. The ID will be
created at runtime, so this CL removes the consumer from
InsetsController while the corresponding source is gone.

However, ImeInsetsSourceConsumer is a special case that it doesn't only
interact with the source and the control, but also InputMethodManager.
So this CL makes the IME consumer always exist.

Bug: 234093736
Test: atest ImeInsetsSourceConsumerTest InsetsAnimationControlImplTest
      InsetsControllerTest InsetsSourceConsumerTest
      WindowInsetsControllerTests
Change-Id: I4ce2663e79618e6634546b86c59092f46e0daef1

Former-commit-id: df821636000efcfcd19225e930134272fc0b6ce6