Make touchXOutsideOfQs require split shade to be enabled
Also renames the method and adds doc. There were 3 calls to
touchXOutsideOfQs and two of the usages were also checking whether
split shade was enabled. Adding the split shade check to the third
fixed b/259591763. Since all usages are paired with the split shade
check, I've moved that into the method and renamed it accordingly.

Fixes: 259591763
Test: manual and atest
Change-Id: I01edaeceac0a04cc1d2d9af22fbdaf88d3a3162f

Former-commit-id: b72e0002ac480ddb44f23f2968f33e0bfe21641c