Do not allow complications when dream overlay is not enabled.
Complications are currently only hosted by the dream overlay. Dependent
logic that listens on the DreamOverlayStateController for changes around
complications should not be triggered when the dream overlay is not
enabled. This changelist addresses this by preventing the
DreamOverlayStateController from accepting complications when the
overlay is not enabled (component wise).

This change also null checks a complication view retrieved from the
view holder before adding it to the layout.

Test: atest DreamOverlayStateControllerTest
Test: atest ComplicationHostViewControllerTest
Fixes: 259859096
Change-Id: Ifa1b4a9da0070a949e422003d3fd4e9ca484a16e

Former-commit-id: 225ce63f0bb27e9bac050cd960ebd7912dd40470