Slice purchase application update notification on locale change
If the locale changes, update the existing notification with the same
information but with resources from the current locale instead of the
previous locale. Use sIntents instead of sSlicePurchaseActivities to
check whether the notification or webview is visible. Don't close
activity on timeout. Fix build errors in tests and add additional tests.

Test: manual verify locale switching, timeout, purchase flow, etc
Test: atest SlicePurchaseBroadcastReceiverTest, SliceStoreActivityTest
Bug: 260262527
Change-Id: I7f0a6a5e32560197e5dc95df00fb59878ad51aa8

Former-commit-id: f8fb10ff1fecc2fd794fc0af8e6f2cf12c74540f