Make AccessCheckingService a SystemService.
And connect it with the existing parts of the system, so that we can
remove all the stub classes created for compilation.

The new implementations of the app op and permission service interface
are now published via LocalServices and available in services-core.

The external states are now actually available, including the package
states being provided by PackageManagerLocal. The callbacks are
temporarily being delivered by the new permission service because only
it has the blocking callbacks at the right timing during package
changes.

A new intialize() function is added for the new app op and permission
services. It is called when the core system services (including ours)
have read the on-disk state and are ready, so that our services can
now retrieve their dependent services and start working. Any call to
the new services before this call can't have meaningful results and
should be no-op'ed in some way.

Bug: 182523293
Test: presubmit
Change-Id: I9c4ec4b7d56ec9d97c4b097cacc7680fa0a08690

Former-commit-id: a8c89fe80740a58094a925f9cf73885f2c759fb4