Attaches source for camera launch when unlocked.
If the double tap power button gesture is detected while the device is
unlocked, we are using a different code path to start the camera app. In
this code path, we were not actually attaching the source to identify
ourselves as the double tap power button System UI gesture.

This CL updates that path to include the source.

Bug: 262193771
Test: manually verified that we are attaching the extra (though it
doesn't fix the camera bug)

Change-Id: If5844a2ceb165303240265f0c0955f94fc8076a0

Former-commit-id: ebe46d7f8a0e71f3efd8060edc2ebb1f1e35ca9f