Revert "Add permissions team to sensorprivacy.proto OWNERS"
This reverts commit 384ac90ff0e26887592af46449fab651f28cb183 [formerly 0f412f7570bca51aa70228a114bb39588b046315].

Reason for revert: 

The CL causes merge conflicts for the updates on AOSP. 

I first tried to cherry pick to aosp. But as "ewol@google.com" is no longer a valid email address, the cherry pick failed.

So I revert this CL and move the correct content to it's AOSP counterpart.
https://android-review.git.corp.google.com/c/platform/frameworks/base/+/2346822

Change-Id: I68f98b963e3ea97d71beb5504d8682ac093edd0f

Former-commit-id: 51e0b300f19966cb40e93bdc4d6f46647075b206