Support fs-verity setup with null/no signature
This is a partial cherry-pick from ag/20131484 (Enable fs-verity to all
APKs on install), where the change to InstallPackageHelper.java is
excluded to avoid the complexity of diverged code base in package
manager.

This is safe because:
1. The function signature does not change.
2. The caller isn't giving it a null signature in AOSP yet, so it won't
   break existing package manager use case.

This partial cherry-pick is desired in order to unblock some work across
internal and AOSP branches.

Bug: 258538225
Test: atest CtsAppSecurityHostTestCases:android.appsecurity.cts.ApkVerityInstallTest
Merged-In: I119e5189603af888dfa1ece2bee9e7635120854b
Change-Id: Ie0001bb74a0b7dffd9f6a3d2a23f5c66f381bea7

Former-commit-id: d163351b15f82dfbcad7c0d6a5e08cee30b94b96