Fix a flicker when short pressing power button.
There is no need to explicitly lock the device on short power button
press if not going to a dream. The default behavior already locks the
device. This fixes a very short flicker when turning off the screen with
the power button.

Also fixes a bug where double-tapping locks the device before showing
the camera.

Bug: 262205573, 262484867
Test: manually by undocking device, tapping the power button, and seeing
there is no flicker before the screen turns off. Also, make sure the
device is locked after the screen turns off.

Change-Id: Iec72119495fb4fd47176094d9a0c33fc312600bd

Former-commit-id: d81ff7be8c2a2ba83b3b9bc7ff487a8a4b194ef8