Remove alt bars from DisplayPolicy
The alt bars were used to decide if the transient bars should be shown
or not, based on their positions on the display. However, the logic of
computing the position was not perfect. It just considered some cases
of gravity, but not all the cases. Instead, this CL uses the insets to
decide which side of display a source window is on.

Also, the alt bar is assigned based on the internal insets type that the
window provides. We won't have the internal types in the furture.

This is a step to remove internal insets types.

Bug: 234093736
Test: atest WindowInsetsControllerTests
Change-Id: Ia9fb857d5be0df57d5d613a033e1e886f04bb479

Former-commit-id: b872b96724a0b9da1524c9cea3ff729ec14d1a48