AudioService-sysUI: sound dose dialogs and interfaces
Define the four sound dose-related events that require displaying
a warning to the user. All warning dialogs are handled by the
CsdWarningDialog class which supports the following behaviors:
- dismissal when shown long enough to user
- triggering an AudioManager volume reduction after a predefined
  time
- not allowing dismissal on volume up
- allowing dismissal on volume down

Implement the triggering of those warnings based on dose
measurements coming from AudioFlinger.

Add property to configure whether CSD is enabled (false
by default). Property needs to be true for any of the
CSD warnings to be displayed.

Bug: 257238734
Test: adb shell dumpsys audio | grep -A 10 "CSD updates"
Change-Id: If0b0756ade7fddb7adfe4dd684ba4583df3ae9ef

Former-commit-id: a5321ff9b470abe71afacd291186c347ab20042b