Fix no FillContext when auth for client suggestions
When the suggestion is coming from the client and does the
authentication, a "No FillContext" error will occur and the
authentication process will be stopped. It is because Autofill does
not create a FillContext for the fill request when requesting client
suggestions. When do authentication for the request, Autofill
appends some data in the intent, like AssistStructure. The structure
is saved in the FillContext. Autofill will try to find FillContext
before authenticating, but can not find, so the authentication breaks.
To fixes, add a FillContext with empty AssistStructure for client
suggestions. And then when doing authentication, Autofill will append
an empty structure in the intent.

Bug: 261817171
Test: atest CtsAutoFillServiceTestCases
Change-Id: I149b7dbf574ed8e0ec96f9e4e8dd7f21a0adc976

Former-commit-id: 9ae0d4b5e271bdca5cd000bac6871fa6e165132f