Populate data sizes in audio egress metrics
The stream and metadata sizes are summed across all HotwordAudioStreams
in the HotwordDetectedResult. Also logged the same byte sizes.

Bug: 261793715
Bug: 244599440
Test: `statsd_testdrive 578` with sample app. The sizes in the metrics
are the same as the ones logged, and the values are in line with what
the sample app is sending in the HotwordDetectedResult.

Change-Id: I48279bba854e3e6badc530fa485bb46c2097ce44
Merged-In: I48279bba854e3e6badc530fa485bb46c2097ce44

Former-commit-id: 10555300ab988a249093d25c4d5a8046c6bf0d4e