Terminate ARC when eARC gets enabled on the RX device
If the RX device has an active ARC connection and eARC disabled,
the user may enable eARC in the RX device's settings, i.e. the Audio
System's settings. In this scenario, the RX will initiate an eARC
connection and should in theory send <Terminate ARC> to terminate the
active ARC connection.

Even if the RX doesn´t send this message, the TX should send the RX a
request to terminate the active ARC connection with a
<Request ARC Termination> message. The TX should send this message when
the HAL reports an EARC_PENDING or EARC_CONNECTED status after the last
status update had been ARC_PENDING.

Test: atest
Bug: 262573690
Change-Id: I8514bcc0858ba6fa744fe2db4c81bdfbed90de22

Former-commit-id: ffa600a03a50a665f49699f6405519c02a3a6ad0