Delegate playSoundEffect calls to VDM for virtual context
When playSound is invoked on AudioManager instance asociated
with context belonging to VirtualDevice with custom device
policy for audio, this change delegates the request to play sound
to VirtualDeviceManager.

With just this change, this effectively disables sound effects for
virtual devices with custom audio policy, in follow-up changes
more options to handle sound effects will be added to
VirtualDeviceManager (custom callback vs. device specific SoundPool).

Bug: 261698699
Bug: 249777386
Test: atest AudioManagerUnitTest
Change-Id: Iba2df745951ed087b8f15a4f210b5b1c7fd936cd

Former-commit-id: fb48c02d3fb53945c136799fdb8e613db1b21beb