Add a checked exception to legacy dexopt code paths.
This exception helps us identify the code in PackageManager that will
eventually be replaced by ART Service. It's propagated to the point
where the code paths need to fork to call ArtManagerLocal instead.

This CL only adds TODO comments at those locations, unless supporting
ART Service is trivial.

This also fixes some cases where Exception is caught. It's an
antipattern because it matches both checked and unchecked exceptions,
so it hides unintended catching of checked exceptions. Instead catch
RuntimeException, along with a specific list of checked exceptions, if
any.

Test: Boot without dalvik.vm.useartservice=true
Test: m FrameworksMockingServicesTests
Bug: 251903639
Change-Id: I53758b8e4b135287e34b412ad1257e5f5fa58273

Former-commit-id: 06ae4cf5320390fffe72d36dbc236f1fd84ce305