Make DreamActivity no history
The DreamActivity is tied to a DreamService. When the DreamActivity is
destroyed, so is the DreamService. If the DreamActivity is
consequently recreated by the WM, this leads to a ClassCastException
in the DreamActivity.onCreate method.

This CL makes the DreamActivity a noHistory activity, which prevents
it from being recreated after having been destroyed once. We also handle
the case where the callback in onCreate is null or of the wrong type and
finish the dream immediately if so.

Bug: 231602954
Test: atest DreamManagerServiceTests
Change-Id: I93637368374d9de90c588e2b0355ae7b66fb4167

Former-commit-id: becfbaf25f5aab659e92e08cc3e0fe6b903044d4