feat(#AlwaysOnMagnifier)!: Supports magnification zooming to 100% [2/2]
There are 2 parts in this feature.
1. support fullscreen magnifier zooming to 100%.
2. support window magnifier zooming to 100%.

This CL is for the 2nd part.

We change the minimum magnifiable scale to 100% in window magnifier, and prevent persistedScale from being 100%.

In WindowMagnificationAnimationController, we add function updateState(). It would be triggered after magnifier enable/disable actions, and update mState by checking the current scale in manager. Therefore, we could assure the controller is in the correct state.

In WindowMagnificationGestureHandler, We add flag mEnabledBeforeDrag to cache whether the magnifier is enabled before entering viewport dragging mode by triple tap and hold gesture. Accordingly, we can decide whether to disable magnifier after dragging end by the flag. Therefore, we can keep magnifier enabled after dragging end if the magnifier is already enabled before dragging start by triple tap and hold gesture.

Besides, we modify some test cases on WindowMagnificationAnimationControllerTest and run atest. Due to Bug/176890994, we remains the @ignore tag.

Bug: 146504199
Test: manually
	atest WindowMagnificationControllerTest
	atest WindowMagnificationAnimationControllerTest
	atest WindowMagnificationGestureHandlerTest
	atest WindowMagnificationManagerTest
Change-Id: Iccdbb42efc34da0dd39b468926259a25c90baf95

Former-commit-id: 9a7f07710518ccd489d51f848eb50c53830cfdf6