[ShellTransit] Reset alpha in buildFinishTransaction
The alpha of a window container can be 0 after it play a fade-out app
transition, (e.g. start/finish activity with overridePendingTransition)
And since there doesn't have another animation layer like legacy
transition do, that window will stay invisible because of alpha=0,
unless somewhere trigger another animation with apply a new alpha on it.
Also it won't able to receive touch event since the alpha is 0.
To make it visible, always reset the alpha to 1 in
buildFinishTransaction.

Bug: 263439418
Test: manual, finish activity with an overridePendingTransition which
apply fade-out animation for the entering one, the activity should be
visible after animation finish.

Change-Id: I5be555e33b76aa799e1f96b3ea340d365bad4c66

Former-commit-id: 265cdc3a689c05ccf11a0e3840beaba94bb425ae