Bind panels with lower priority flags
When showing a panel in device controls, bind to it (but with lower
priority flags). That way, the application is kept in memory and results
in shorter latency when relaunching the panel with the TaskView.

Use BIND_NOT_PERCEPTIBLE so it will be at lower priority from SystemUI.
However, don't use WAIVE_PRIORITY, as by itself, it will kill the app
once the Task is finished in the device controls panel.

Test: manual
Test: atest com.android.systemui.controls
Fixes: 263790347

Change-Id: Ia99ade528bfdbfe2f67ecfe78f687f18df8b6cfd

Former-commit-id: 7173a3fa5763b006d7b28127ec97e05e358fbb8a