Don't register for UDFPS long press in STATE_DOZE
We only want to register for the udfps longpress when the
device is in AoD. This includes the AOD_PAUSED state in
case the user immediately attempts to authenticate with UDFPS
right after taking out of their pocket. We allow UDFPS longpress
to trigger in that state even though UI hasn't shown yet.

Test: atest DozeSensorsTest DozeTriggerstest
Fixes: 245657872
Change-Id: Iaecb0e37deed6ad2ba10eb8cbda098bf6443c01f

Former-commit-id: f3b7a9aff4b6c7b7fcb740c75b33045d17d655c2