Add methods to mark some constraints as optional.
These methods will let apps mark battery-not-low, charging,
and/or idle as optional (but preferred) instead of just as a
hard requirement. This gives JobScheduler more freedom to defer
the job to more ideal times if it can without confusing developers
further.

For now, the methods are hidden because there's a little bit more work
to be done on the communication and documentation.

Bug: 236261941
Test: atest CtsJobSchedulerTestCases:FlexibilityConstraintTest
Test: atest CtsJobSchedulerTestCases:JobInfoTest
Test: atest CtsJobSchedulerTestCases:JobSchedulingTest
Test: atest frameworks/base/services/tests/mockingservicestests/src/com/android/server/job
Test: atest frameworks/base/services/tests/servicestests/src/com/android/server/job
Change-Id: Ia27947b3a40f59d0278698e3d12a6b656b18aa16

Former-commit-id: eb72459a8994180ddc1c9ee05d15ac31413ee2ea