RESTRICT AUTOMERGE Update FP listening state on keyguardGoingAway
And handleFinishedGoingToSleep. This goes back to earlier behavior when
we used to always update the fingerprint listening states when
keyguardGoingAway and finishedGoingToSleep signals were sent.

A bug was introduced when ensuring face auth would only STOP when
keyguardGoingAway & finishedGoingToSleep signals were sent. Since
face and fingerprint are different, we separate their logic in this CL.

Test: atest KeyguardUpdateMonitorTest
Test: manually see that fingerprint listening state starts quickly
after pressing the power button to screen off
Fixes: 261025550

Change-Id: If5e5d2132640d77a5eaa5765d932c6509f8d7bc8

Former-commit-id: e21879df4e2b8663000ab9f08d51b01bc005eff2