Add DISABLE_REASON_APP_ONLY and parameter to navigate to app
With this functionality, apps can disable specific routes
in the output switcher so that they are only available in
the app. A usecase for this is customizing the way the
routing session is joined.

This change also adds a ComponentName which the system can
use to navigate to the app when the user taps on a route
disabled using DISABLE_REASON_APP_ONLY.

Test: atest MediaRouter2HostSideTest
Bug: 241888071
Bug: 235352899
Change-Id: Ia291dfb9ad9a940b10f2c1b759dc1ee205df6cc2

Former-commit-id: 02406603087560ee083f56f24e7062ab333bf2ab