Override alpha to 0f if we're no longer interactive.
This is a safe way to ensure the surface is immediately
no longer visible, until the screen off lifecycle event
arrives and ends up canceling the unlock sequence and
sleeping the device. WM resets the surface alpha after
the animation ends, so this should be safe.

Since the surface alpha is usually close to 0f when this
condition is met, it's not usually perceptible that we
jump-cut to 0f.

Test: press the power button while also triggering an unlock
Test: atest KeyguardUnlockAnimationControllerTest
Bug: 260094933
Change-Id: I15a9decf1dcfac0b60b3f4e97a2e470098961743

Former-commit-id: 767467732e3c237293a1c2734e7feb1040f66070