Unsuppress USI low battery notifications when new stylus used.
This handles an edge case where a stylus with a different
vendor/product ID is used after a previous stylus has received a
low battery notification, and that notification has been dismissed.
Unsuppressing the notification means that if the new stylus's
battery is also low, a new notification will be sent.

Note that this means if a stylus has the same vendor/product ID
as the previous one, there is no way to differentiate them.

Bug: 262689155
Test: atest StylusUsiPowerStartableTest
Change-Id: Ia8a661a131e241216ecf9aaff89e36cb40860f91

Former-commit-id: c644d2b9685d4d7ce14766b3521655d39394ba3d