[SB Refactor] Connect old pipeline visibility/tints to new mobile icons.
The _new_ pipeline has the source of truth for whether the mobile icons
*should* be visible. But, the _old_ view system has the source of truth
for whether the mobile icons *can* be visible (i.e. if there's enough
space for it).

The _old_ view system also has the source of truth for icon tint.

This CL hooks up the old system's knowledge of visibilities and tints to
the new pipeline view.

Bug: 238425913
Test: manual: verify the mobile icon tints correctly
Test: manual: add a lot of subscriptions and verify that the
subscriptions on the left eventually shot a dot or hide completely
Test: manual: verify turning flag 611 on/off correctly shows/hides the
icon debug coloring
Test: atest ModernStatusBarMobileViewTest

Change-Id: I94d58d468eab8370b49ffbc97967570991ddd3d9

Former-commit-id: f7055e6d607841b1c2b5f61946831d875a01a6f7