Trigger unfold animation without treshold while on keyguard
When folding, the fold animation triggers at different angles based on what's on the screen. Before this cl, it was triggering regardless of the angle only while on launcher, and after a certain threshold in all other cases.

While on keyguard, the angle depended by the app that was in foreground. (e.g. if calculator was open, and device locked, the threshold was 60 degrees).

After this change, while the keyguard is visible, there is no threshold.

It has been needed to create the intermediate UnfoldKeyguardVisibilityListener as DeviceFoldStateProvider is in the global dagger scope, while KeyguardStateController in the sysui subcomponent (so, injecting KeyguardStateController directly in DeviceFoldStateProvider is not possible).

From launcher, it is expected we don't have the information whether keyguard is visible (as we don't need it).

Bug: 261313897
Test: DeviceFoldStateProviderTest & manual
Change-Id: I07f133e96b16572448754ce2009f55e9e316b136

Former-commit-id: aa55dda145ce0d206ac4f481ab5061582b764a92