Per-app compat treatment for overrideToPortrait.
Camera framework consumers will have the option to opt in or out to
the rotate and crop fix for foldables via a manifest property.
This will allow apps which correctly handle the landscape sensor
to obtain the full FOV of the camera.

Test: Tested opt-out in Camera2Basic sample, opt-in in CTSVerifier.
Bug: 263965273
Change-Id: I190963c4f8dfba29dab7570397c33f44cab702e3
Merged-in: I190963c4f8dfba29dab7570397c33f44cab702e3

Former-commit-id: eb361cec47438fda430b2a9fee522ba01a6c0cf2