Fix transition animations for launch-from-staged operation
This change makes it so that a smoother animation will play when the user cancels the splitscreen flow by tapping the staged app to launch the app in fullscreen.

Previously, we did not support the launch of a solo app into fullscreen during the splitscreen flow. When this feature was implemented in ag/20432050, the fade-in animation would subtly fail (and the solo task would pop in with no animation at all) because the single-app animation case was not fully supported yet. This change adds a single-app-launch case in startTasksWithLegacyTransition() and startIntentAndTaskWithLegacyTransition(), allowing the animation to be applied in the same way as in the splitscreen case.

Fixes: 257513449
Test: Manual
Change-Id: Ieadf8cecf39d36d4252152ba55c6d9f84634281a

Former-commit-id: 65514324d99772a7206acbff03320329fc2bf035