Do not postpone draw transaction
Otherwise if the operations are merged to mPostDrawTransaction
when the buffer is requested to be synced, the transaction order
may be unexpected. Then the layer may appear to be frozen and
unable to handle subsequent operations, including removal.

The activity window may have sync handler because it is
transforming by SeamlessRotator.

Reference changes:
[1]: I00606c0ed250a10e1651c403665f39be52e9a89e
[2]: Id89d039328bd9b091d4ed78a1616bc78d1bd83fa

Since the reparenting for preserved surface is gone, it no longer
needs a separated WSA#mPostDrawTransaction.

The mPostDrawTransaction was used for
[3]: Ie91da3c78eaf8fb52a21ae49ed190b42d92d0083
[4]: I14c2f7603345ad89f0af4db48150b05d8ada602a

Bug: 263009910
Test: atest WindowStateTests#testApplyWithNextDraw
Test: Disable auto rotation.
      Launch Youtube (disable PiP) and expand to landscape.
      Take screenshot and press edit (it should show in portrait).
      The screen is not hang when exiting the edit activity.
Change-Id: I02eaa8a16153a12b4a605600e3c75beb38c6f5e0

Former-commit-id: ecae534c9dc7dc9dfa9231d8d010b60a1a464bb2