[Sb refactor] Add a reusable method to TableLogBufferFactory
The majority of uses of LogBuffer and TableLogBuffer will be set up in
the dagger graph, as the controllers logging to it are typically known
at compile time.

In the case of mobile, we are lazily creating the log buffers, and want
to reference them by name. This CL adds a cache to
TableLogBufferFactory, which is opt-in to avoid wasting memory caching
every single log buffer. They are all already retained by DumpManager
which throws an exception if a reregister is caught.

Test: MobileConnectionsRepositoryTest
Fixes: 264257095
Change-Id: I0e93e4317eb8108288983ffa01ccbc092e1e242e

Former-commit-id: a79001aab495ecc03e5ae576c0456e2b07d54138