Take DynamicCodeLogger out of DexManager.
Before this change, DynamicCodeLogger was inside DexManager. However,
once we switch to ART Services implementation, DexManager will no longer
be used unless the user opts out ART Services. Therefore, we need to
take DynamicCodeLogger out to keep it functioning.

This is a no-op change. The DynamicCodeLogger - ART Services integration
will be done in a separate CL.

Bug: 263165518
Test: Presubmit
Change-Id: Ia19ea397f9db8e9f0095da47b98631ddec11fa17

Former-commit-id: a7cb4d15d90da5d7317da6025e1979cea7ac529a