Ignore power presses from lift and biometric wake
WAKE_REASON_GESTURE previously used to include both WAKE_REASON_LIFT,
WAKE_REASON_TAP and WAKE_REASON_BIOMETRIC. Since ag/20794573, these
gestures are not not all encompassed by WAKE_REASON_GESTURE. This CL
ensures the power suppression logic is still used for wake reason lift
and biometric.

We intetionally include biometric in case biometrics are related
to the power button.

We intentionally leave out tap to wake since it's not
necessary to gate a power button press due to a wake up from a tap
on the screen.

Test: Press the power button while performing lift to wake gesture,
observe that the device does not go back to sleep
Fixes: 264553999

Change-Id: Ia74d30fd808d0a49937f483575b53533b0f1fce7

Former-commit-id: c1e2804be14d9510575e7af745e122351f9077c7