Remove RecordingCanvas's reliance on Skia's private SkTemplates.h
SkAutoTMalloc was previously used through some transitive inclusion of
SkTemplates.h. Skia is trying to move these private types into an
internal namespace to address leakage, so this usage must be removed
first.

AutoTMalloc has the same API surface as SkAutoTMalloc, but is not the
same due to SkAutoTMalloc delegating most of its work to other private
Skia dependencies and had to be rewritten.

AutoTMalloc is declared in utils/AutoMalloc.h since there are other
"AutoMalloc-like" variants that will be needed in HWUI to replace other
Skia types.

Changes in include ordering are from required clang reformatting.

Bug: 264403827
Test: existing presubmits (should be a drop-in replacement)
Change-Id: I1763ce0a643ba3e2f7be58e860dbdf08e3ba07bc

Former-commit-id: d7e45e4fb521ae9b5d71e07f621353ddde0687da