Allow a PendingIntent's creator to revoke BAL privilege.
Currently if the creator of a PendingIntent is allowed to start a background activity the PendingIntent has the same privilege even if started by another process in the background that does not have this permission. This change allows creators to opt out of this behavior. Nothing changes if the creator does not actively opt out.

Test: atest  com.android.server.wm.ActivityStarterTests
Bug: 261780465
Change-Id: I29db44d4fb07fa4a178a0d2e5822f655cfa56c39

Former-commit-id: 64c8239c8541fe03c9811caf7931c1c574fa7c80