Remove Dream Overlay when Dream Window detaches.
This changelist removes the dream overlay window when the associated
dream window is removed. This follows closely to the attach behavior,
where the dream overlay window is added after the dream window is
attached. Previously, the dream overlay window was removed during
onDestroy. Deferring removal to this later point in the lifecycle
allows for the window to be incorrectly acted upon after the associated
dream is gone.

This changelist also implements a default implementation of
DreamOverlayService#onWakeUp which immediately runs the callback. It
also addresses an issue where service connection consumers were not
cleared in the overlay service connection.

Fixes: 261913526
Test: atest DreamOverlayServiceTest
Test: atest DreamOverlayTest
Change-Id: I4c80b585a484c20a8282fca75bd1908b494ce928

Former-commit-id: 0aa0f4acb29b4b61d66aabc609d0be3ced06aeeb