RESTRICT AUTOMERGE Center align the lock screen wallpaper
CP for UDC initial release where lockscreen live wallpaper is not enabled, and later we need to revert this for udc-qpr-dev.

On smaller display of a multi display device the lock screen wallpaper
should be showing the center area of the wallpaper. Check the display id
to find out which display the lock screen wallpaper is on, the check
needs to happened when we are adjusting the bounds to get the new
display id.

Bug: 263932616
Bug: 289079608
Test: set the lock screen wallpaper and observe the result
Change-Id: I32b93084a9e716a06e032f83a8cf625c25d3e5e5
(cherry picked from commit 40070be04f46e80f9ee3114da738c4b76192a298 [formerly c979b749192a609d630173f944053e805ae58c11])

Former-commit-id: c9e25a5d5fc8ec938ec028461446a8d6186ebb2e