[Chooser/ResolverActivity] Add option to show preview area even if there are no apps in the tab
Add a protected method that could be overriden
to change the default behavior that hides the
content preview area when there are not items
for the current user.

This is needed for the partial screen sharing
feature as we want to show the recents app selector
no matter if we have apps available for the
current profile or not. Recent app selector is
displayed in content preview area.

Test: atest com.android.internal.app.ChooserActivityTest
Test: atest com.android.internal.app.ResolverActivityTest
Test: manual test with overidden method to check that the preview is displayed when no apps available
Bug: 233348916
Change-Id: Iecc7ea5b15cb6cd2d864bd7adc84ab2f9f445fca

Former-commit-id: f51b8efaed4ef71716cb64a20177ba43975a2103