Make ScreenshotRequest handle hardware bitmap conversion
Currently, ScreenshotHelper provides utility functions to convert
between a hardware bitmap and a parcelable form for sending between
processes, but the originating code must handle the conversion itself.

This change moves the conversion so that it occurs seamlessly
inside of the ScreenshotRequest class, and refactors the inputs
to ScreenshotHelper slightly so that the ScreenshotRequest data class
is what gets passed in.

Bug: 264457397
Test: atest
Change-Id: I1041bbcfdc5f5ece6e98bd0844017d4f3dc4e9f3

Former-commit-id: 414f66073a649ec76f26a1ab1717959eece5b1b2