Make mutable PendingIntent explicit
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the PendingIntent here was only
destined to the test app/to the app, so it was made explicit. Reviewers,
please call out if this is not the case.

Bug: 236704164
Bug: 229362273
Test: atest IntentSenderTest
Test: atest RemoteViewsTest
Test: atest NotificationTest
Test: atest SuggestionTest
Test: atest NotificationStressTests
Test: atest android.content.pm.PackageSessionTests
Change-Id: I6721b52b63c406724855f5e3824281ef6965fa75

Former-commit-id: f929d5a5ca9a2a8d1f641714bfcbad4b84d21bbb