Make mutable PendingIntent explicit
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the PendingIntent here was only
destined to the test app/to the app, so it was made explicit. Reviewers,
please call out if this is not the case.

Bug: 236704164
Bug: 229362273
Test: atest ShortcutManagerTest9
Test: atest ShortcutManagerTest8
Test: atest RecoverySnapshotListenersStorageTest
Test: atest RecoverableKeyStoreManagerTest
Change-Id: Ib6bc41c6985e90e6b7d42a5b509686996c9c5425

Former-commit-id: edf533081313337a37321d5215998d68f4ccae18