Transitions - Update for canceled state
When transitions are canceled a new one begins, existing animations
could be left in an incorrect state (position, alpha), so make sure to
end them where they should be. Also, if a transition is started
halfway, make sure to emit events to inform the ViewModel where it
should be at that point in time.

Also, prevent oscillating events for dreaming by adding in a slight
delay.

Bug: 260637747
Test: atest KeyguardScenariosTest

Change-Id: I03c9f25175f745e5eca93ab1ba74cd7f7005ee28

Former-commit-id: caedd59764edf4316cbbd16f02935df174e082bd