Tweak how we skip transitions for first activity
The first home activity is inserted into a task and then started from
ActivityStarter by Task.startActivityLocked.

Previously we checked for the first ever activity by checking for
isHome (=true) and not hasActivity (=true). This was intended to come
back false if there was no activity before this one. However the
activity is already created by this point so we wrongly treated this
first launch the same as any other launch.

Another way of checking for this condition is to do what the comment
says: see if getActivityBelow(x) exists, if not we have nothing to
animate in on top of.

Bug: 258882804
Change-Id: I6ffa281bad6bf5ffcf738e281c987f3a21a7ace1

Former-commit-id: 2dde31ef86a49748784b3a0d2539de34e0546dff