Move early-wake to transition start
Any binder to SF is actually pretty slow esp. on low-end
devices. Early-wake just needs to be set "before the animation",
so move it to transition-start and merge it with the global
transaction (which gets applied no-matter what in `animation`).
This actually bring it more in-line with legacy (where early-wake
is more likely to be set before the first frame) and it remove
the latency of making the blocking binder to set it in its own
transaction.

Bug: 258909157
Test: record trace when launching app (or run the benchmark
      in the bug) and check the "animating" track.
Change-Id: I31e0cf4ea0644c327a99c4dc62047e6e465c853c

Former-commit-id: c40d1edfe623430cf552d39ac5e66225c26eaec2