BinaryTransparencyService: Fix printPackageInstallationInfo
It seems that APEX_PRELOAD_LOCATION_ERROR is consistently returned
as the value of original APEX location when calling
`getOriginalApexPreinstalledLocation` when preloaded APEXs are
compressed.

While this change does not address the root issue of why that is
the case, it prevents the error value to be used as a legitimate
path to compute digests for by guarding the control flow based on
sane values of apex install location.

Bug: 264925790
Test: Manual.
	adb shell cmd transparency get apex_info -v
Change-Id: I924ba889cb6d11a874cf06ca8692f8320b86b588

Former-commit-id: 00001111031ca5206f06d5d0e4ad877f3ff595da