[Sb refactor] Add the default network's connectivity to the view model
Part of the criteria of whether or not to show the RAT indicator is
whether or not `mobile` is a connected transport. The way the old
pipeline does this is by keeping track of the default network
capabilities (via `ConnectivityManager`) and letting those be known by
all mobile connections.

The equivalent in the new pipeline is to expose the default network
capabilities via the MobileIconsInteractor, and pass it down to each
individual icon interactor.

Finally, the reason this is sensible to do in the new pipeline is that
we can track when the data subscription changes and thus we can avoid
icon flickering (see following CL)

Test: MobileIconViewModelTest
Test: MobileIconInteractorTest
Test: MobileIconsInteractorTest
Bug: 264683083
Bug: 238425913
Change-Id: Ib3a4f14ba3cb403f342090b7394d271309a3a7f5

Former-commit-id: 5ceb454c8e8465ddd6df1e046f46719b3bb1702d