Make mutable PendingIntent explicit
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the PendingIntent here was only
destined to the test app/to the app, so it was made explicit. Reviewers,
please call out if this is not the case.

Bug: 236704164
Test: atest AccessibilityManagerServiceTest
Test: atest SystemActionPerformerTest
Change-Id: I6fb8ff75c05d2bc631586fb1e14ea6820c213807

Former-commit-id: f7bbf68f083758e0dca188ef91d7bdaebd9a9aba