Revert "Enable tethering namespace accessible in Android native level."
This reverts commit 78aa3f7c43df0e47bf7ad96150dc1351939c6ff6 [formerly 32bd9d6d0563482d98624f046a979622b1618f50].

Reason for revert: Original commit was for native experiment setup.
But SEPolicy was not updated and sync to native level always fail in T devices.
This CL will clean up incomplete native experiment setup.
Even if we decide that we want to use flags from native code in the Tethering module,
we will do so using a new namespace, tethering_u_plus_native.

Test: m
Bug: 264947218
Change-Id: If32fe0b237dcb213886efb72430ea7575ab64530

Former-commit-id: 78ef4714f58c5d205e6c1775bcd78f2ac4c8df88