Make mutable PendingIntent explicit
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the PendingIntent here was only
destined to the test app/to the app, so it was made explicit. Reviewers,
please call out if this is not the case.

Bug: 236704164
Bug: 229362273
Test: atest OverlayManagerPerfTest
Test: atest CtsPackageInstallTestCases
Test: atest NotificationManagerServiceTest
Change-Id: I56cc8c1852caffadae424e08d7719f58779bab11

Former-commit-id: 1efd9a0405c6b8d5ec0b00f327b588b9eb69ac0a