Add a SystemApi constant for TIME_MANAGER_SERVICE
The TimeManager API was added in S. There is no actual "time manager
service" (time and time zone functionality are split into different
services the system server), but the APIs pretend there is one.

The existing constant name "TIME_MANAGER" fails lint checks for
inconsistent naming, i.e. lacking "_SERVICE, so it has to be renamed to
"TIME_MANAGER_SERVICE". Lint then complains about it needing to be
TIME_SERVICE, but I have drawn the line there and suppressed it.

Bug: 265148320
Test: compile only
Change-Id: I2a94e01520b56f49d6761d0aeba4e1b879c2a64d

Former-commit-id: 023accd8636ccff68cd4f3277f5afe3e6d4e34a2