Make implicit mutable PendingIntents immutable
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the PendingIntents here are not
mutated, so they can be immutable.

Reviewers, please call out if this is not the case.

Bug: 236704164
Bug: 229362273
Test: make -j
Test: make checkapi
Test: adb shell cmd package list libraries |\
      grep com.android.media.remotedisplay
            And check the com.android.media.remotedisplay library
Test: atest NotificationAssistantServiceTest
Test: m StatusBarTest
Change-Id: Ibf1cbe6b5636ef3b02cfcf24d44ed3d4a4357d82

Former-commit-id: 17394f3dbf671b6b0abe385bcdfc65cb889826c9