Add a persistence mechanism for panels' packages
This will allow to persist the packages of apps that provide device
controls panels after they are authorized (opted in) by the user.

Packages that are in the seeding list (config_controlsPreferredPackages)
will be auto-added, but only when creating the set of authorized package
for the first time (per user).

Test: atest AuthorizedPanelsRepositoryImplTest
Fixes: 264414289

Change-Id: Ic23e6172158bba3cf960e81919978ec62da79150

Former-commit-id: e192a439af9a5025f7be515e39078b584c73461c