Added ADB exception case for Background Installed Apps detection.
This is a temporary fix as it is not ideal to have initiatingPackageName
be null for ADB installed package events. There is a bug cut to
PackageManager team to address this at b/265203007. Will follow with a
long term fix once this bug is resolved.

PatchSet 5: Updated to use initiatingPackageName as that is immutable.

Change-Id: If80bb72a6cc02c73004d4e0665537255f727d0a0
Bug: 264568994
Test: atest

Former-commit-id: eeefd04f560b653669683726531bdc286e626a7b