OomAdjuster: Give SDK sandbox the priority of the app it belongs to.
SDK sandbox processes are bound to by system_server, but system_server
binds on behalf of a regular application process. An earlier changed
stored this "attributed" process in the AppBindRecord; here, retrieve
it, and use that to compute the priority of the sandbox instead.

Additionally, keep track of started SDK sandboxes per UID, so that we
can retrieve them when computing the downstream dependencies of
applications in collectReachableProcessesLocked().

Bug: 253399592
Test: atest SdkSandboxLifecycleHostTest
Test: evaluate OOM adj scores
Change-Id: Ib0696c55e7c0c16c3ba2d19d861098bd2e3ebd21

Former-commit-id: 2a3daaa3b12fe428ec79ca8a201e829b6370073c