Step 1 of Removing Ids from Flags.
With this change, we start relying more directly on the string names
that are now associated with flags. The ids persist so that we can
look up existing overrides and push them into the new system.

After a couple of weeks, the plan will be to remove the ids entirely.

Bug: 265188950
Test: manually built before and after cl to ensure overrides persist.
Change-Id: I0faac671b43a0d24262e78ccdb4e23e44f73eeea

Former-commit-id: cd1af6a4d14ee3f9fa4948ae229981a5d7bfc0aa