Fix missing KeyguardStatusView after cancelling fps login
FPS login will trigger keyguard fading away. If that is cancelled by
pressing the side button, then the keyguard will instead transition
to AOD. Depending on the timing of the transition, this can leave
keyguard fading away as set, causing any subsequent animation that
doesn't reset it correctly to hide the KeyguardStatusView.

Fixes: 262112441
Test: Manually verified that race doesn't leave keyguard in bad state
Change-Id: Id20a300edfa52433642747dbcdb1ecceeffadba8

Former-commit-id: 3df40e98cb524e39e58103af0ee307aa48ac4d98