Do not go to dream on power button press when screen is off.
This is essentially a revert of a previous change to start dreaming on
power button press when the device is docked and the screen is off. We
now just use the existing (default) behavior of turning on the device to
the lockscreen.

Bug: 265349281
Test: manually by docking device, turning the screen off, and then
tapping the power button to turn the screen back on. The device should
show the lockscreen and not attempt to start dreaming immediately.

Change-Id: I25bb8bc5b720fd6cc1b1ac9219c8c94eecf8232a

Former-commit-id: f37ffbacc096d97b051915d3547d963a443919d1