Publish dream state to OverviewProxyService.
This change pipes the current dream state to OverviewProxyService,
similar to how doze information is shared. A condition has been
added to monitor the dream state, and a callback relays this
information to the StatusBarStateController. A new CoreStartable
maintains the monitor for this interaction.

Test: atest StatusBarStateControllerImplTest#testSetDreamState_invokesCallback
Bug: 265755700
Change-Id: Ic927a7e55864453da026c14cd1c87e6a7cf22489
Merged-In: Ic927a7e55864453da026c14cd1c87e6a7cf22489

Former-commit-id: 58b6e0dfa296ac3babadfafeacbe6d9406f82e28