Optimize app compaction system reduce page faults
This patch allows for amortizing the memory churn in compaction
over time by reusing the vector and its vmas as well as sending
a preassigned buffer to fill maps. Effectively, after a compaction
run we usually see no new allocations which also reduces the page
fault cost and the allocation cost.

Bug: 243976314
Test: am compact full <processname> <uid>

Change-Id: I36e96284b72b7b03d75171058def23b783f6c5cd

Former-commit-id: 2b7a8e18832b64fab6d985d4f3292755cc6ca1d4