Fix warning log in DialogLaunchAnimator
This CL fixes a warning log that is now logged every time a dialog is
launched from a View, instead of being logged only when launching from a
View that does not implement LaunchableView. This log was broken by the
refactoring of http://ag/20139844, which abstracted away the View we
launch from into a Controller.

Bug: 230830644
Test: Looked at logs during dialog launch
Change-Id: I9e7ee5bff4dbea109a8e53137a8613d5e5d4ef37

Former-commit-id: f05f5f9324a1641bc0ba347a686e2a809b5efbd8