Adds userId to onTopActivityChanged callback in VirtualDeviceManager
Adding the userId will allow onTopActivityChanged to be associated
with a userId when the top activity changes in a stream via using
the VirtualDeviceManager and allows the API user to know which
profile is active. While the userId can be found through other
means, the indirect path can cause a race condition with poor user
experience.

Bug: 264474826
Test: atest CtsVirtualDevicesTestCases
Change-Id: Ib00e1589a2c7b545d499d65e1d7ff9c12f0b4654

Former-commit-id: d849e0006f069919f8f6fc3d71fac7c5c405211c