Direct connection API for runtime sensors.
The type of direct channel and highest supported rate level are
reflected in the sensor flags. Only MemoryFile channels are supported
for virtual device sensors, HardwareBuffer ones are not.

New sensor callback methods inform the VirtualDevice owner when a new
direct sensor channel has been (un-)registered or configured for a
sensor, along with the writable shared memory region and rate level.

Bug: 266042170
Test: atest SensorControllerTest
Test: atest VirtualSensorConfigTest

Change-Id: I239bdf448ce0ff1412a0bcd0a4654d70a08b3077

Former-commit-id: 5c4d5c86b50f011e38c99fa78ff8720cbbfd3a64