Make Notification long clickable
Notification aka ExpandableNotificationRow handle long click but it was not marked as long clickable. That confused test and that would have a possible impact on accessibility.
We implemented unit test which checks whether the notification is long clickable or not to verify the behavior and fixed the issue by marking Notification long clickable.

Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.notification.NotificationGuts#testExpandAndCollapseGuts

Bug: b/264948201
Change-Id: I34248b9e207c93c6e987f70603e8b3725c7c1ab1

Former-commit-id: 4574b5fd595b866ed8ff413df8a76ca81c44bb00