Hack: Clear the mDozeAmount=1 override when switching to SHADE to fix notifications going invisible
Another bug has come up where the (AmbientState) dozeAmount==1 triggers notifications to disappear when expansion==1.  The NotificationWakeUpCoordinator state and logs indicate that this state persists because of the bypass override. Digging in, this appears to be the result of the bypass state changing in a way that class never expected.  This change attempts to track when the doze amount is overridden to 1, and allow that override to be reset when the keyguard state changes away from KEYGUARD.

Additionally, when unfolding from the LS, the bypass state change wasn't triggering the clear, so before this change even if the eventual unlock fixed the notification state, the inside lockscreen could be left in the wrong state.  The bypass state listener which checks and conditionally clears the override is enough to fix that bug as well.

Fixes: 265626464
Test: manual repro of notifications disappearing: 1) tap to wake, 2) trigger face auth, 3) unfold quickly, 4) check shade (and logs) for symptom.
Test: manual repro of bypass state when unfolded: 1) tap to wake, 2) do not show face 3) unfold, 4) see only icons (no notifications) on inside screen.
Change-Id: I8dacae4aeb0dbb6c744643acd1b0122fd4f3eb22

Former-commit-id: 826eefe0ee7d844080c779cdab16b755dc497983