docs(non linear font scaling): deprecate TypedValue.scaledDensity and add warning on Configuration.fontScale
These fields are no longer useful for calculating font size, since font
size is now derived using a non-linear curve.

There is no reason anyone should need to use TypedValue.scaledDensity,
so that is deprecated.

However, it still could be useful to know the user's font scaling
preference for other reasons (e.g. choosing a different layout that fits
larger font sizes better), so we leave that one intact, but with a
warning clarifying it should not be used in hardcoded equations.

Bug: b/237558231

Test: nope
Change-Id: I5f61492025d1ac66aee302ca0790cebc422d4f55

Former-commit-id: c64189b0c424689d876ed5fcefcc1943077822d7