Add support to make the refresh rate used when the device has HDR enabled configurable for each display.
This is a step forward to prepare DisplayModeDirector to load the config
per display when hdr is enabled for the device to keep the changes
contained. As a followup, we will add support to also load the config from the
respective display config file.

Bug: 264965782
Bug: 265126371
Test: atest DisplayModeDirectorTest
Merged-In: Ia76f732274eea08ad83e8479820fb096eb134dee
Change-Id: Ia76f732274eea08ad83e8479820fb096eb134dee

Former-commit-id: 9365343aad6658435f8b51f9c1ef0d81994f30bf