Provide Context to InputManager and deprecate getInstance
Apps will obtain InputManager through the Context#getSystemService call.
Historically, InputManager was obtained through the static getInstance()
method by the framework code, so no Context was provided to
InputManager.

In this CL, we deprecate the getInstance() method, and provide the
Context as a late-initialized variable when the static InputManager
instance is first obtained through the Context.

Bug: 266013036
Test: presubmit
Change-Id: Ie35baae9d6b5576e7fb8047e0f5dd3de97835f0d

Former-commit-id: 7c1d972e0c2b1738093ae801fef04e31f60dd522