Lock screen long-press to open wallpaper picker.
Implements a long-press gesture on the lock screen that opens the
wallpaper picker (WPP) screen (after properly unlocking the device).

- Respects device state
  - Only works on the lock screen - does not work when unlocked, quick
    or quick-quick settings is shown, etc.
  - Lets other lock screen UI components receive and handle touch
    (including smart space, unlock/UDFPS, shortcuts, shade, and
    bouncer)
- Flag gated - this is a brand new flag as we're not sure whether we
  want to roll this out and it can be rolled out berore the WPP revamp
  project
- Logging included - logs long-press popup impressions and clicks, to
  help inform launch/no-launch decisions and identify falsing challenges

Fix: 265987364
Test: manually verified that it matches the UX spec: does not work
unless locked and showing the lock screen un-occluded (tested occlusion
with the camera app). Does not work if quick settings or quick-quick
settings is visible. Does not work if unlocked with or without the shade
being shown.
Test: manually verified that smart space clicks, notification clicks and
long presses, UDFPS, and shortcut buttons all still work as advertised.
Test: unit tests included

Change-Id: I600f3e60a03b58691ffd66cddb1be6ff0c80b3ca

Former-commit-id: 73260bd178c326f31bbd8e999e87aa4748f285c7