Propagate ancellation signal to cred man service
If developer cancels the get/create request, we propagate
to the remote cancellation signal and clear the session.
This change includes manually checking before sending back a
response to the develoeper, before invoking the UI and before processing
the repsonse of UI, and then propagating the cancellation signal to the
providers.

Could follow up with adding a listener that proactively cancels the
session in addition to the hook points checked manually.

Test: Built & deployed locally
Bug: 264944428

Change-Id: Ia4c0c4d0e758433400d4f086b482fdb273120968

Former-commit-id: 53dacd6f2d1cbf914a1679304b44143cd8943b7b