Remove CanvasCompare.
As part of renderscript deprecation/removal, of the two pieces of
renderscript in this app, one already had a Java alternative but the
heatmap code didn't. Rather than reimplement that, though, it appears
that this project has been obsoleted by CtsUiRenderingTestCases. So
let's remove it instead. (Even if we're wrong, this still gives us more
time to work out what, if anything, is actually needed.)

Bug: http://b/255624710
Test: treehugger
Change-Id: I7cf64a28bd3c7edea061fafc5e06d4134cc79b05

Former-commit-id: 2ae77029fda816d029b4bfe2620bd4f01022aab0