Fix incorrect return value from getDeviceVolumeBehavior
getDeviceVolumeBehavior may currently return an incorrect volume
behavior if the current behavior is not persisted in settings, and
setDeviceVolumeBehavior was called in the past.

With this change, getDeviceVolumeBehavior will always read the internal
state instead of settings to get the current volume behavior.

This affects a System API used by privileged apps, which may behave
incorrectly as a result.

Bug: 266566375
Test: atest AbsoluteVolumeBehaviorTest
Change-Id: I9fc3699741abe1123706ebf980cdf052566d4d31

Former-commit-id: c80109db1a017d3fa6699cc38f5114dcb14d7464