Move ViewCapture On/Off controls to QuickSettings Tile.
Rather than use a feature flag for this feature, the on/off state will
be stored as a system setting and will be changed via a QuickSettings
tile.

Bug: b/264452057
Test: Verified that the new QuickSettings tile doesn't crash via normal
interactions (pressing, long-pressing, etc.). Also verified that
ViewCapture is turned on when the QuickSettings tile is in the enabled
state and is turned off when it is in the disabled state.

Change-Id: I19b26c279d627ccc8a242536aac83de07c824db9

Former-commit-id: 4181dd38fc806ff14b591c2af2f42ad6daaeb6c5