Keep keyguard visible if we're launching an occluding activity over a going-away keyguard.
This happens when bypass is enabled - the occluding activity
launches over the keyguard visually, but the keyguard is
going away since it's being dismissed by the launch. We
weren't checking for this case, so we were allowing keyguard
to be hidden, showing the shade UI.

Additionally, we need to keep the wallpaper around to prevent
the keyguard UI from showing over a black backdrop during
this animation.

Fixes: 266630755
Test: launch occluding activity over unlocked lockscreen w/ bypass enabled
Change-Id: I5a7df7df5aa5339276470529e7164475c6e99e13

Former-commit-id: 88885b0e68fc1327ebf20ce70d36df82112b36c7