CSD: Persist csd and recover after boot or audioserver crash
Using the Settings.Global to persist any new CSD entry on the
AudioService side. Reason for global is because the standard is defined
for a PMP and not user. When the system boots we recover the previously
persisted CSD. After a crash in audioserver we apply the AudioService
cached values to the restarted audioserver. This will include the
recovery of the csd value and records which lead to that value.
New settings should not be backed up since they are determinedi for each
device independently.

Test: logs and dumpsys after audioserver kill and reboot
Bug: 264254900
Change-Id: Ie0c7b400d466981d198145d0ca0fc9c9e29eca63

Former-commit-id: 774b6d568320900cfc0517337d2d6f9a6e7ab92c