Add clock-type to choose how timestamps are provided when profiling
Add an option to am start-activity and am profile to be able to choose
the clock source for the timestamps. The clock source can be wall /
thread-cpu / dual. wall provides the wall clock timestamps which is
fast. thread-cpu provides the time the actual time thread was on cpu.
This is slow and should be only used when actually required. Dual
provides both these timestamps.

Bug: 259258187
Test: am start-activity --start-profiler --clock-type wall / thread-cpu
Change-Id: I719395d673c337f577ba5245814338e35d69c372
Merged-In: I23a01ff91e1a5c7e8a39339d06d0bfad8a6c4c37

Former-commit-id: c58be09d9273485c54d6a16defc42d9f26182b73