Enable color threshold editing via SystemProperties
Add SystemProperties for when Wallpaper Colors decides to suggest dark
text (threshold) and how much noise (amount of dark pixels) to factor in for the wallpaper color decision. This will allow for editing of these properties via the commandline, so that we can find
the best threshold for dark/light wallpapers.

Bug: 202758428
Test: manual
Change-Id: I6e2415e15b41040c5250c46b0f0dbe77e6224f36

Former-commit-id: 2ea3f16fe74d1ca4435ec2fadcdd2901364897c9