Allow inactive recommendation card and set timeout
When the flag is enabled, the recommendation card can be inactive,
rather than being removed immediately once it is swiped away. If the
smartspace update is marked as "PERIODIC_TRIGGER", it will not make the
carousel active. The card can still be removed manually with the long
press menu. In addition recommendation cards will now be removed automatically
when its expiry time is reached rather than waiting for an empty update.

Smartspace updates not marked as periodic will have the same behavior as
before (make carousel active, showing recent media first if it exists)

Test: atest com.android.systemui.media.controls
Test: manual with test build + smartspace flags
Bug: 264690420
Change-Id: I78d365a2df8b7654041bd425745343232cd19bc4

Former-commit-id: 243ed5881985e5fc5b4ee13cbcdad2a6878f69a9