Rename Bubble#isClearable to isDismissable
We have previously changed (ag/16576767) how we set the Bubble#isClearable field, but
we didn't follow up on renaming this field. BubbleEntity#isClearable is set with the value of NotificationEntity#isDismissable and is used for dismissability checks.

The CL renames this field to follow the same convention as in regular Notifications.

Fixes: 267156792
Test: atest BubbleXmlHelperTest
Change-Id: I47a7cac87ff3e48eefc0a3872c06715834af8e67

Former-commit-id: c764ab22a449ab5f90c691f3d77af1cc9be5935a