Update transition tracing for FaaS usecase
For FaaS, we don't really care about all the intermediate states of a
transition we only want to get information about which window and
surfaces were part of the transition and how they should have changed.

These changes provide the data for FaaS in a format that is more
convinient for FaaS, and avoid additional processing on the FaaS side.

It also means that we can be more memory efficient when tracing the data
we need for FaaS, which allows us to enable always on tracing to provide
meaningful data about transition that FaaS can then use to run on traces
for dogfood devices.

Test: N/A
Bug: 267484886
Change-Id: I4d7aae1414e45a8170b71143e6bfa804064bbfce

Former-commit-id: 3705aa4151aa0b93d46f96d36ede5dae7c568c05