Allow shortcuts to be pinned by unbundled chooser
Originally ShareSheet was part of android system, so shortcuts pinned by
ShareSheet are pinned by the system process. Since we are moving
ShareSheet into a separate module, this means shortcuts should not be
pinned by the system process, but by unbundled chooser process, when
that module is enabled.

This CL change the package name that pins the shortcut to corresponding
package in respect to whether intent resolver module is enabled.

Bug: 258072921
Test: Steps to verify
1. Open Chrome, go to any website.
2. Click on the search box, click share to share the URL
3. Click on more to bring up system ShareSheet.
4. Long-press on a direct sharing shortcut, then pin the shortcut.
5. Verify the direct sharing shortcut is pinned in the share sheet.
6. adb shell pm enable --user 0 com.android.intentresolver/.ChooserActivity
7. repeat 1-5

Change-Id: Id7c5b9d7cb17aa0dc8e5293bba9de29f8197e182

Former-commit-id: 81d0988803cb2b6b6f93fb06ffd5156f2c990b79