AudioService: fix stream aliasing update
When updating the stream aliases table, mStreamVolumeAlias
should point to a copy of the default alias table it's starting
from, not its reference.
This fixes an issue where changing the aliasing between ring
and notification would work when unaliasing, but not when
aliasing, as the contents of the alias table had been overwritten
already.

Bug: 265076070
Test: atest android.media.audio.cts.AudioManagerTest#testStreamTypeAliasChange
Change-Id: I74bc7fdf054ef295876c8a5503111c14d084e8d2

Former-commit-id: e620bc2a8e5f236b4d57a0a525a81eda2578425b