Keep existing focus during transient-launch.
If the user hasn't moved enough, the touch event is supposed to
go through the gesture-bar to the app. Once the user has moved
enough, though, the inputs should go to the launcher.

In legacy, touches automatically went to the live-tile because
the launcher was launch-behind (and thus not focused) but an
input consumer had to be added once the gesture started to take
events away from the app.

Shell is the opposite, launcher is on top so it would get focus
automatically; however, we need to add logic for the app to
hold-onto its focus until the gesture starts. This CL adds a rule
where transient-launches do not automatically take focus. If it
wants focus, it can make an explicit focus request.

Bug: 266660298
Test: SettingsActivityUiAutomatorTest#unblockAppDialogTest
      CloseImeWindowToHomeTest
Change-Id: I5ab01dc56dd4ae8bd40d5035ea985f9d562ac866

Former-commit-id: 3e62d4f2c4bed7109932d7ec279e4c9d22e2a97f