[1-time permissions] Use internal api to check proc states
We need to check the proc state and the binder method has a filter that
is affected by a bug that keeps a killed a proces in the "pending top"
list. Using the internal api isn't affected by this filter and also is
more correct for inprocess calls.

Test: Install test app that requests permission and will exit
        immediately on granting, observe permission is no longer
	indefinitely held.
Bug: 254736794
Merged-In: I30579090c803b231fd750abbc4ad645805f7ece2
Change-Id: I30579090c803b231fd750abbc4ad645805f7ece2

Former-commit-id: e836611f3057cf9eae589a34a39fe80d0a9145f3