Use consistent naming for sharing broadcaster's identity
Android 14 added APIs to allow a broadcasting app to share its
identity with receivers, but a consistent naming convention was
not used in the internal APIs. This commit updates the internal
APIs to use a standard "sending" prefix for the uid and package
of the broadcasting app.

This commit also resolves the following issues unrelated to the consistent naming convention:

- If the broadcast is being delivered to the original app that sent it, skip granting app visibility since it is not required.
- When the result is delivered to the app that sent the old ordered broadcast, use the information from the oldRecord to determine if identity should be shared.

Bug: 259743961
Test: atest ShareIdentityTest
Change-Id: I80783636e4f85ee66d253f56ee83d4b5faac4158

Former-commit-id: 7277f26fde08080b9d8bc520996d72f9d3bad396