Avoid changing resolution by display mode when animating
There was a logic to use default refresh rate when animating (e.g.
default is high refresh rate, but app requested low refresh rate).
But since display mode can also affect display size, if the mode
triggers the size change when animating, the animation may be
interrupted or canceled. So it is better to keep the current mode,
and the mode will be updated on the window traversal after the
animation is finished.

Also do not set display properties when display is seamless
rotating, otherwise there will be obvious flickering because
there won't be screenshot to cover the resolution change.

Fixes: 238771010
Bug: 255850221
Bug: 267365057

Test: RefreshRatePolicyTest#testAnimatingAppOverridePreferredModeId
Change-Id: I5b3ddbc8278a95b4c5239812b2c4a293775c740a

Former-commit-id: e6c20eee8d26a42c59f6ffe63dda997e8bf53fc6