Update the default backup user after onUserUnlocking if necessary
On the first ever boot of a HSUM device, UserManager will create the main user after BackupManagerService is created. If that happens the default user will stay as 0 until the device is rebooted. To prevent this we need to update the default user as soon as it's guaranteed that the main user is created (if one is to be created), which is when the first user unlocks.

Bug: 266098768
Test: atest BackupManagerServiceTest
  atest BackupFrameworksServicesRoboTests
  manually by doing a factory reset and then checking backup is active for the main user via `adb shell bmgr --user 10 activate`
Change-Id: Ic14e7357b468c27d206a57f9e7148d918f747c1d

Former-commit-id: c9892805d16df1ed5abba424e4df8e6e39f83422