Fix heads-up notification being sometimes clipped when pulsing
The NSSL top padding is aligned with the clock top. The clock position moves throughout
time to avoid pixel burn-in and sometimes the position is above NSSL bounds, which makes the HUN be clipped.

The solution is to account for burn-in vertical shift when calculating NSSL position.
This means smartspace with clock will not always be aligned with HUN but it's agreed to be the best solution.

Also small renaming of burn-in related fields.

Test: KeyguardClockPositionAlgorithmTest.java
Test: Manually on-device with simulating biggest burn-in shift up
Fixes: 260084123
Change-Id: I211b15affef1c84d0b8524134b3af02894faf536

Former-commit-id: 5c7cee2b3a12b6bc003abd8fca25a840bdf83bd6