Fix javadoc for the uid paramter.
The original doc says uids are user IDs, which is correct in the sense that they are Linux user IDs, but super confusing because in 99% of the cases we use "user ID" to refer to Android users instead of Linux users. This has caused confusion to developers and we should fix that by simply saying it's a UID which we use to refer to the Linux UID we create for sandboxing every app.

Change-Id: Ic24849fc443649b0656827c5a3f271e0688dab1a
Test: presubmit

Former-commit-id: e3a9966a0c1e4bcc686161436953cffa3da96038