Don't send transitions to Shell until finished booting
While booting, the display isn't enabled yet so transition
animations aren't visible. Additionally, there can be
problems during this time because not all parts of the
system are fully initialized yet causing animations to not
run properly and potentially trigger ANRs.

This CL "localizes" all transitions that start during the
boot phase. This should be faster and prevents coupling
WM and Shell processes while the system is already very
busy.

Bug: 267439165
Test: SystemAppCheck#system_app_crash
Change-Id: If6625e57fcd2246835eaa5dc0155465631e8019c

Former-commit-id: 84db17e81d9f1875b24cfdbbb435a0311bae751a