Update rest of DPC tests to use a holder.
Since the tests can cover multiple DPCs, rather than having copies of
all the mocked items live in the global space, the last patch introduced
a new holder to scope them to the DPC they're related to. That pattern
seems to work well, so apply it to the main DPC under test as well.

Bug: 265793751

Test: atest DisplayPowerControllerTest
Change-Id: If722af7408432a4b80d79e20a0007f695c03180c

Former-commit-id: d6afb3186b8bd790b97622adbbd4c14482226720