Add InputSettings for input related settings
The original intent of InputManager was to wrap a remote service;
however, over time some functions were added that don't match the
original intent. This is the first step to remove some of those
extra functions.

This CL adds InputSettings, a new class that encapsulates the
functions that expose and modify input related settings.

Bug: 267758905
Test: compiles, and should pass presubmit
Change-Id: Iedd625aacc34407412cc05b048e25081bb90950b

Former-commit-id: 7ec4828bcf598baf3595ecdb6598acf463048f77