Added changes for utilising a common feature flag
This CL aims at updating feature flag used for platform storage of PIN
length to make it the same as the one we use in Settings. This is to
make sure we use a single feature flag for Auto PIN Confirmation
feature.

Test: Tested by building and flashing on local. And modifying the
feature flag value using mobdog and checking if the PIN length is
getting stored or not.
Bug: b/262935308

Change-Id: I61304bb9e9518222267f7810d9b2509fcc02bc67

Former-commit-id: 2d500436f26477224ae6745aafa19ae97e8c723b