Changed DPMS and UMS to explicitly start a profile.
Traditionally, when a profile was started as background user, the
profile would be "visible" (i.e., allowed to launch activities), as
long as its parent is also visible (for example, if the parent is the
system user, the profile would initially be visible, but it would
become invisible when the current user is switched).

On Android U, the concet of "background" was explicitly divided into
"background invisible" and "background visible", as the latter can be
used to start full users visible on passenger displayes (on
automotive).

So, to keep the profile behaviour consistent - and to allow profiles
to be started as "background invisible" when needed - they should be
explicitly started with `startProfile()` (instead of
`startUserInBackground()`).

Test: atest UserControllerTest FrameworksServicesTests:UserManagerServiceTest
Test: manual verification with TestDpc (set a managed profile and
      used QS to turn it on / off)
Fixes: 268045186

Change-Id: I268d47940c23db4ef0e04770f70a2b59d86f9d29

Former-commit-id: 750ef177b1635863d4c7a71e4537ca6113d5b12f