Make all copied URIs shareable
We can only display thumbnails for images, but the sharesheet can
handle other types of URIs. Add the share button for these types
of copied data.

Also switches to not show UI if the clipData is null, and just show a toast if the clipboard contents have no items. (These cases should be approximately unreachable.)

Bug: 268057213
Fix: 268057213
Test: atest
Change-Id: I4c1e118c38c48c8ea8fd0d89ee426d7d961460af

Former-commit-id: e9e8071989257f98a90e4c9921da0e6d22c6154e