[wm] Drop no focused window ANR request if the activity is stopped
If the activity is in stopped state it will not have a visible
window on screen. In this case drop any ANR request which tries to
blame the app if it's focused without a window. This can happen on
some products which do not have a focused system window when the
display is turned off.

Test: turn of display, send keyevent
Bug: 236763025

Change-Id: I773bcd9b3448654705cdad224ea4426f9545f3cf

Former-commit-id: 9c09066ff6acac8c01eb6d61b723b54ee6779426