Fix requestUnbufferedDispatch didn't work before window attached
Before window attached to view, the parent of root view would be null
cause the unbuffered request can't be updated to the `ViewRootImpl`.

This will update the unbuffered request when the root view set to its
`ViewRootImpl`, so we could make sure the unbuffered request could be
consistent in root view and its ViewRootImpl.

Bug: 268105812
Test: manual
Change-Id: Ifd9cf0969229daeac73f8fb95c6a9fe8c81a985f

Former-commit-id: b2a805df9f44e8acf537cb9f80d06114b6edf778