Stop AndroidFrameworkContextUserId on UserManager
For older APIs, UserManager uses the method getContextUserIfAppropriate() to get the context user (if a newer API level) or the calling user (if an older API level). We had to gate on API level so that we could convert old API usage to the new context-user-based style.

In this cl, we recognize this method as a legit context-user-style method, so that we can stop getting AndroidFrameworkContextUserId lint warnings on UserManager.

Change-Id: I95ee95645314ab9105a48a800de06d9b8c7ecb85
Fixes: 268129776
Test: atest error_prone_android_framework_test

Former-commit-id: c80cb2eb9d4831049c3abbb35cfb55752028c589