AudioService: Fix setting LeAudio volume after VoIP call
When mode changes to NORMAL but the USAGE_VOICE_COMMUNICATION is still
present in the playback streams the mode change would trigger volume
update but would still use STREAM_VOICE_CALL volume level and the music
content will be resumed with voice call volume, possibly with higher
and potentially harmful volume level.

This change relies on mode change to get stream type from the mode
and therefor restores the right volume right away, without relying on
the onPlaybackConfigChange() which seems to be happening to late while
the loud music with the conversation volume level already reached the
listener.

Bug: 247659576
Tag: #feature
Test: manually testing with VoIP call while streaming media
Change-Id: Ie2225b7e92e0d3f7146e9266caf8bda81eddad03

Former-commit-id: d461d4762a3f7b03ad062cc97b6a4084fe55a635