[Sb refactor] Ignore old pipeline callbacks to removeIcon*
StatusBarIconController was already ignoring calls to set icon when the
new pipeline was enabled, but was still allowing calls to `removeIcon`
and `removeAllIconsForSlot` externally. Therefore, it was possible for
the new pipeline to add an icon for a mobile connection during system
startup, only to have it later removed by `StatusBarSignalPolicy`, which
later came on line and tried to remove existing subscriptions.

This change ignores external calls to remove icons if the slot is
something controlled under the new pipeline.

Fixes: 266960478
Test: StatusBarIconControllerTest
Change-Id: I570fe9ea2e8c3cf09c89a100a68dce73e00febf7

Former-commit-id: 754e1951fbe028d61330c325086a17dc4e79b435