Resume installation after permitting installing unknown apps.
When a user allows an app to install unknown apps, resume the
installation. PIA checked for an incorrect result code from the settings
activity, which is now fixed. Instead, we check if the client is allowed
to install unknown apps before resuming.
PIA will finish itself if the user returns from the install unknown apps
settings page without toggling the permission.

Bug: 267992205
Test: Manual. Followed the repro steps in the mentioned bug.
Change-Id: I17eca83c7fbfb3a610fe9cd009f3e61e4723a651

Former-commit-id: a310dd355f880b06c9ec913ca4b4df838b57fd5e