Fix race between startNew and Core-request transition
If WMCore requests a transition while a startNew from Shell
is "in flight", Shell could be expecting startNew to be
processed first while WM is actually processing its own
requested transition first. This can lead to incorrect
behavior or an NPE.

To resolve this, make sure the WM request transition is
the first non-started transition in Shell's records (basically
syncronizing Shell to Core's transition order).

Bug: 265861362
Test: atest ShellTransitionTests#testTransitionOrderMatchesCore
Change-Id: I62a25335f270a021a06dd4309e692e8812d43332

Former-commit-id: bc9f4bb3efadd61189238a89eec8cda0cd77fa29