Ensure transition is ready after applying display change
Previously, the ready state of display change (e.g. rotation) relied
on TRANSACT_EFFECTS_LIFECYCLE to call resumeFocusedTasksTopActivities
that will call executeAppTransition if all activities are already
resumed. But if not, the transition will wait until timeout with
message "timed-out because not ready".

And because sendNewConfiguration already ensures configuration,
visibility of activities and requests traversal, only need to ensure
it is ready then the transition animation can start when the sync
group is finished.

Bug: 266711229
Test: atest DisplayContentTests#testShellTransitRotation
Change-Id: I9c441f97e5e0a0bf66c1fa73fae45755890ff463

Former-commit-id: 85afcfa54d7e86899e4a12bbc6917938236df626