More eagerly collect isSimpleUserSwitcher.
This is a blind attempt to fix the attached bug. The hypothesis is that
there's a race condition between when this is queried by the large
screen bouncer user switching drop-down and the coroutine job that
hydrates this flow with values.

The approach in this CL is to sacrifice some startup performance to make
the value available as early as possible as to avoid the race condition.

Bug: 256698234
Test: all existing user-related unit and integration tests pass locally
Test: manually verified on both small and large screen devices that the
popup and full-screen user switcher, respectively, show up
Test: manually verified, on large screen, that the bouncer
user-switching drop-down shows up and works

Change-Id: I431dbace23f8472a296b374739729a9867efe8ea

Former-commit-id: 210b958f161437798caab1a5becc9e24d06c79f9