Increase max shader cache size
Bumps maxTotalSize (max size of the entire cache) to align with
the previous bump to maxValueSize (max size of an single entry in the
cache) in I1be3da0458c4b64f7082a2e1bc0d08b0824bbeba, which was discussed
on b/190348616. I chose 4 MiB here to keep the new values linearly
proportionate to their original values.

Note: this entire cache is stored in both RAM and flash. If 4 MiB is too
much, we could alternatively set maxTotalSize to 2 MiB and reduce
maxValueSize to 1 MiB (if maintaining the original proportions is
important).

Bug: 190348616
Test: builds (with new static_assert) and passes hwui_unit_tests
Change-Id: Ib101d1c432f3547483f52e64fad81bf0a1e20f07

Former-commit-id: f50917cd0afbffaf5006563a25a3db5d19f26449