[Media TTT] Don't use an animated-vector for the loading spinner.
See bug for more context. tl;dr: AnimatedVectorDrawables will pause
their animation when opening the shade (or in other scenarios), so we
need to use a different kind of drawable as a workaround. This CL just
uses a static drawable and animates its rotation using an
`ObjectAnimator`.

Bug: 243983980
Test: `adb shell cmd statusbar media-ttt-chip-sender MyTablet
TRANSFER_TO_RECEIVER_TRIGGERED` -> see new loading spinner. Pull down
the shade and verify the loading spinner keeps spinning. Plug the device
in to charge to see the charging animation and verify the loading
spinner keeps spinning.
Test: _TRIGGERD -> _SUCCEEDED -> Undo ==> verify that the new loading
spinner for the new TRIGGERED state is spinning
Test: atest ChipbarCoordinatorTest

Change-Id: Ic5701bbb241b919211463df6eb576f810ec2ca22

Former-commit-id: 287842965a2c05a261bce332d975737df16aaab4